+++ apps/teams/adapter.py
@@ -1,5 +1,3 @@
-from allauth.socialaccount.adapter import DefaultSocialAccountAdapter
-from django.core.exceptions import PermissionDenied
 from django.urls import reverse
 
 from apps.users.adapter import AccountAdapter
@@ -29,16 +27,3 @@ class AcceptInvitationAdapter(AccountAdapter):
             except Invitation.DoesNotExist:
                 pass
         return super().get_login_redirect_url(request)
-
-
-class SsoAccountAdapter(DefaultSocialAccountAdapter):
-    def pre_social_login(self, request, sociallogin):
-        # Retrieve the initial email from the session
-        initial_email = request.session.pop("initial_login_email", None)
-
-        if initial_email:
-            provider_email = sociallogin.user.email
-
-            # Check if the provider's email matches the initial email
-            if provider_email.lower() != initial_email.lower():
-                raise PermissionDenied("The authenticated email does not match the one provided.")
